Split the navigation with tests
