+++ core/test/navigation.spec.js
@@ -161,89 +161,6 @@ describe('Navigation', function() {
     });
   });
 
-  describe('getChildren()', () => {
-    const nodeWithChildren = {
-      children: [{ name: 'children1' }, { name: 'children2' }]
-    };
-    const nodeWithChildrenProvider = {
-      children: () => {
-        return [{ name: 'children' }];
-      },
-      context: [{ name: 'context' }]
-    };
-    const nodeWithChildrenProviderError = {
-      children: () => {
-        throw new Error('Error!');
-      }
-    };
-    it('should not fail if arguments are undefined', async () => {
-      const children = await navigation.getChildren(undefined, undefined);
-      expect(children.length).to.equal(0);
-    });
-    it("should return empty array if it doesn't have children", async () => {
-      const aNode = {};
-      const children = await navigation.getChildren(aNode, undefined);
-      expect(children.length).to.equal(0);
-    });
-    it('should return nodes children and bind them if children are provided', async () => {
-      const children = await navigation.getChildren(
-        nodeWithChildren,
-        undefined
-      );
-      expect(children).to.equal(nodeWithChildren.children);
-    });
-    it('should return nodes children and bind them if children provider is provided', async () => {
-      const children = await navigation.getChildren(
-        nodeWithChildrenProvider,
-        undefined
-      );
-      expect(children).to.equal(nodeWithChildrenProvider.children);
-    });
-    it('should not fail if children provider throws an error', async () => {
-      const children = await navigation.getChildren(
-        nodeWithChildrenProviderError,
-        undefined
-      );
-      expect(children).to.be.undefined;
-    });
-    it('should return children using provied context and bind them', async () => {
-      const children = await navigation.getChildren(
-        nodeWithChildrenProvider,
-        'context'
-      );
-      expect(children).to.equal(nodeWithChildrenProvider.children);
-    });
-    it('uses navigationPermissionChecker and returns correct amount of children', async () => {
-      //given
-      window.Luigi = {
-        config: {
-          navigation: {
-            nodeAccessibilityResolver: (
-              nodeToCheckPermissionFor,
-              currentNode,
-              currentContext
-            ) => {
-              if (nodeToCheckPermissionFor.constraints) {
-                return nodeToCheckPermissionFor.constraints === 'other_scope';
-              }
-              return true;
-            }
-          }
-        }
-      };
-
-      const nodeWithChildren = {
-        label: 'someNode',
-        children: [
-          { label: 'child1', constraints: 'some_scope' },
-          { label: 'child2' }
-        ]
-      };
-      const children = await navigation.getChildren(nodeWithChildren);
-      expect(children.length).to.equal(1);
-      expect(children[0].label).to.equal('child2');
-    });
-  });
   describe('findMatchingNode()', () => {
     it('substitutes dynamic path', () => {
       // given
